backend/requirements.txt
# FastAPI Backend Requirements for Million-Dollar AI Music Generation System
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CORS and Security
fastapi-cors>=0.0.6
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
motor>=3.3.0  # MongoDB async driver

# AI and ML
torch>=2.1.0
transformers>=4.35.0
diffusers>=0.24.0
librosa>=0.10.1
essentia>=2.1b6.dev1034

# Audio Processing
soundfile>=0.12.1
scipy>=1.11.0
numpy>=1.24.0
mido>=1.3.0
pretty_midi>=0.2.10
pyaudio>=0.2.11
aubio>=0.4.9

# Music Theory and Analysis
music21>=9.1.0
mingus>=0.6.1
pyfluidsynth>=1.3.2

# Cloud and Infrastructure
boto3>=1.34.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.10.0

# Monitoring and Logging
structlog>=23.2.0
prometheus-client>=0.19.0
sentry-sdk>=1.38.0

# Background Tasks and Queues
celery>=5.3.0
dramatiq>=1.15.0

# WebSockets for Real-time Updates
websockets>=12.0

# Caching
aioredis>=2.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
pre-commit>=3.6.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.1

# File Processing
Pillow>=10.1.0
python-magic>=0.4.27

# API Documentation
sphinx>=7.2.0
sphinx-autoapi>=3.0.0

# Performance
aiofiles>=23.2.1
asyncpg>=0.29.0
orjson>=3.9.0

# Blockchain Integration (for NFT minting and royalties)
web3>=6.12.0
ethereum>=2.3.2

# Payment Processing
stripe>=7.8.0
paypal-checkout-serversdk>=1.0.1

# Email and Notifications
sendgrid>=6.11.0
twilio>=8.11.0

# Analytics and Metrics
mixpanel>=4.10.0
amplitude-analytics>=1.2.0
